+++ sentinel/sentinel_ncs2.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+source /home/pi/openvino/bin/setupvars.sh
+/home/pi/.virtualenvs/openvino/bin/python -u /home/pi/sentinel/sentinel.py